Version 0.1.21.2
Dynamic Sprite testing. Currently have it working with a test new label
type called newLabelType, which uses a PlaneGeometry. Now able to create
a multiline sprite with text of any length, detect whether the mouse is
over a path inside the canvas (workaround to canvases having to be
square), and call the PlaneGeomentry "Intersected" only when the mouse
is within the path on the canvas.